fix: memberservice의 메소드 이동으로 인한 테스트 오류 수정
